Show option for saving changes to the server only when customization->forcesave is true.

+++ apps/documenteditor/main/app/view/FileMenuPanels.js
@@ -132,10 +132,10 @@ define([
                     '<td class="left"><label id="fms-lbl-autosave"><%= scope.textAutoSave %></label></td>',
                     '<td class="right"><span id="fms-chb-autosave" /></td>',
                 '</tr>','<tr class="divider autosave"></tr>',
-                '<tr class="edit">',
+                '<tr class="forcesave">',
                     '<td class="left"><label id="fms-lbl-forcesave"><%= scope.textForceSave %></label></td>',
                     '<td class="right"><span id="fms-chb-forcesave" /></td>',
-                '</tr>','<tr class="divider edit"></tr>',
+                '</tr>','<tr class="divider forcesave"></tr>',
             /** coauthoring begin **/
                 '<tr class="coauth changes">',
                     '<td class="left"><label><%= scope.strCoAuthMode %></label></td>',
@@ -318,6 +318,7 @@ define([
             this.mode = mode;
             $('tr.edit', this.el)[mode.isEdit?'show':'hide']();
             $('tr.autosave', this.el)[mode.isEdit ? 'show' : 'hide']();
+            $('tr.forcesave', this.el)[mode.canForcesave ? 'show' : 'hide']();
             if (this.mode.isDesktopApp && this.mode.isOffline) {
                 this.chAutosave.setCaption(this.strAutoRecover);
                 this.lblAutosave.text(this.textAutoRecover);
@@ -372,9 +373,11 @@ define([
                 value = 0;
             this.chAutosave.setValue(fast_coauth || (value===null ? this.mode.canCoAuthoring : parseInt(value) == 1));
 
-            value = Common.localStorage.getItem("de-settings-forcesave");
-            value = (value===null) ? (this.mode.customization && this.mode.customization.forcesave) : (parseInt(value)==1);
-            this.chForcesave.setValue(value);
+            if (this.mode.canForcesave) {
+                value = Common.localStorage.getItem("de-settings-forcesave");
+                value = (value === null) ? this.mode.canForcesave : (parseInt(value) == 1);
+                this.chForcesave.setValue(value);
+            }
 
             value = Common.localStorage.getItem("de-settings-spellcheck");
             this.chSpell.setValue(value===null || parseInt(value) == 1);
@@ -396,7 +399,8 @@ define([
             Common.localStorage.setItem("de-settings-fontrender", this.cmbFontRender.getValue());
             Common.localStorage.setItem("de-settings-unit", this.cmbUnit.getValue());
             Common.localStorage.setItem("de-settings-autosave", this.chAutosave.isChecked() ? 1 : 0);
-            Common.localStorage.setItem("de-settings-forcesave", this.chForcesave.isChecked() ? 1 : 0);
+            if (this.mode.canForcesave)
+                Common.localStorage.setItem("de-settings-forcesave", this.chForcesave.isChecked() ? 1 : 0);
             Common.localStorage.setItem("de-settings-spellcheck", this.chSpell.isChecked() ? 1 : 0);
             Common.localStorage.setItem("de-settings-showsnaplines", this.chAlignGuides.isChecked() ? 1 : 0);
             Common.localStorage.save();
