+++ apps/presentationeditor/main/app/view/FileMenuPanels.js
@@ -122,10 +122,10 @@ define([
                     '<td class="left"><label id="fms-lbl-autosave"><%= scope.textAutoSave %></label></td>',
                     '<td class="right"><span id="fms-chb-autosave" /></td>',
                 '</tr>','<tr class="divider autosave"></tr>',
-                '<tr class="edit">',
+                '<tr class="forcesave">',
                     '<td class="left"><label id="fms-lbl-forcesave"><%= scope.textForceSave %></label></td>',
                     '<td class="right"><span id="fms-chb-forcesave" /></td>',
-                '</tr>','<tr class="divider edit"></tr>',
+                '</tr>','<tr class="divider forcesave"></tr>',
                 /** coauthoring begin **/
                 '<tr class="coauth changes">',
                     '<td class="left"><label><%= scope.strCoAuthMode %></label></td>',
@@ -268,6 +268,7 @@ define([
                 this.chAutosave.setCaption(this.strAutoRecover);
                 this.lblAutosave.text(this.textAutoRecover);
             }
+            $('tr.forcesave', this.el)[mode.canForcesave ? 'show' : 'hide']();
             /** coauthoring begin **/
             $('tr.coauth.changes', this.el)[mode.isEdit && !mode.isOffline && mode.canCoAuthoring ? 'show' : 'hide']();
             /** coauthoring end **/
@@ -304,9 +305,11 @@ define([
                 value = 0;
             this.chAutosave.setValue(fast_coauth || (value===null ? this.mode.canCoAuthoring : parseInt(value) == 1));
 
-            value = Common.localStorage.getItem("pe-settings-forcesave");
-            value = (value===null) ? (this.mode.customization && this.mode.customization.forcesave) : (parseInt(value)==1);
-            this.chForcesave.setValue(value);
+            if (this.mode.canForcesave) {
+                value = Common.localStorage.getItem("pe-settings-forcesave");
+                value = (value === null) ? this.mode.canForcesave : (parseInt(value) == 1);
+                this.chForcesave.setValue(value);
+            }
 
             value = Common.localStorage.getItem("pe-settings-showsnaplines");
             this.chAlignGuides.setValue(value===null || parseInt(value) == 1);
@@ -322,7 +325,8 @@ define([
             /** coauthoring end **/
             Common.localStorage.setItem("pe-settings-unit", this.cmbUnit.getValue());
             Common.localStorage.setItem("pe-settings-autosave", this.chAutosave.isChecked() ? 1 : 0);
-            Common.localStorage.setItem("pe-settings-forcesave", this.chForcesave.isChecked() ? 1 : 0);
+            if (this.mode.canForcesave)
+                Common.localStorage.setItem("pe-settings-forcesave", this.chForcesave.isChecked() ? 1 : 0);
             Common.localStorage.setItem("pe-settings-showsnaplines", this.chAlignGuides.isChecked() ? 1 : 0);
             Common.localStorage.save();
